Move application commands to separate file
The validate and update functions are now in commands.ts. This file
should be used for the functions called by any of the program
commands from the entry point. Added test file for commands.
