+++ src/index.ts
@@ -1,12 +1,8 @@
 #!/usr/bin/env node
 
-import util from 'util';
-import path from 'path';
-import { exec } from 'child_process';
 import { program, Option } from 'commander';
-import fs from 'fs-extra';
-
-import { config, validate } from './utils';
+import { validate, update } from './commands';
+import { config } from './utils';
 
 main().catch(error => {
   console.log(`Encountered an unexpected error: ${error}`);
@@ -35,23 +31,4 @@ async function main() {
     .action(update);
 
   program.parseAsync(process.argv);
-}
-
-async function update() {
-  try {
-    // check if the repo exists. if not, clone it. if it exists, fetch updates.
-    if (!fs.existsSync(path.join(config.SCHEMA_REPO_FOLDER, '.git'))) {
-      await util.promisify(exec)(
-        `git clone ${config.SCHEMA_REPO_URL} "${config.SCHEMA_REPO_FOLDER}"`
-      );
-      console.log('Retrieved schemas.');
-    } else {
-      await util.promisify(exec)(
-        `git -C "${config.SCHEMA_REPO_FOLDER}" checkout master && git -C "${config.SCHEMA_REPO_FOLDER}" pull --no-rebase -t`
-      );
-      console.log('Updated schemas.');
-    }
-  } catch (error) {
-    console.log(`Error when updating available schemas: ${error}`);
-  }
 }
